requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-telegram-bot==13.7
six==1.17.0
urllib3==1.26.18 